refactor(frontend): reorganize project structure
- Move globals.css from src/app/ to src/styles/
- Consolidate all React hooks into src/hooks/
  - Move useNetworkStatus from lib/hooks to hooks/
  - Move useAudioPlayer and useTrackPreloader from lib/audio to hooks/
  - Rename use-pwa to useServiceWorker for clarity
- Reorganize utils directory structure
  - Move cn function to lib/utils/cn.ts
  - Create lib/utils/index.ts for unified exports
  - Update all imports (18+ files)
- Move api-config.ts from lib/api to lib/constants
- Remove legacy files
  - Delete dashboard-nav-refactored-example.tsx
  - Remove src/app directory (Next.js remnant)

All changes validated with lint, tsc, tests, and build
