+++ frontend/src/components/layouts/adaptive-layout.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import * as React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/cn";
 import { useDashboardLayout } from "./use-dashboard-layout";
 
 type AdaptiveLayoutContextValue = {
