+++ frontend/src/components/ui/badge.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { type VariantProps } from "class-variance-authority";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/cn";
 import { badgeVariants } from "./badge.variants";
 
 export interface BadgeProps
