+++ frontend/src/components/ui/empty-state.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { VStack } from "./stack";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/cn";
 
 interface EmptyStateProps extends React.HTMLAttributes<HTMLDivElement> {
   /**
