+++ frontend/src/components/ui/stack.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/cn";
 
 interface StackProps extends React.HTMLAttributes<HTMLDivElement> {
   /**
