+++ frontend/src/hooks/useTrackPreloader.ts
