+++ frontend/src/lib/utils/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Utility Functions Index
+ * Re-exports all utility functions for convenient imports
+ */
+
+// Tailwind CSS class name utilities
+export { cn } from './cn';
+
+// Duration formatting
+export { formatDuration } from './format-duration';
+
+// Hash utilities
+export { calculateFileHash, calculateBufferHash } from './hash';
