+++ frontend/src/main.tsx
@@ -2,7 +2,7 @@ import React, { Suspense, lazy } from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import './app/globals.css';
+import './styles/globals.css';
 
 // Import i18n initialization
 import './locales/init';
