+++ frontend/src/pages/LibraryDetailPage.tsx
@@ -12,11 +12,11 @@ import { useLocale } from "@/locales/use-locale";
 import { formatDuration } from "@/lib/utils/format-duration";
 import { AddSongToPlaylistForm } from "@/components/features/libraries/add-song-to-playlist-form";
 import { DeleteSongButton } from "@/components/features/libraries/delete-song-button";
-import { useAudioPlayer } from "@/lib/audio/useAudioPlayer";
+import { useAudioPlayer } from "@/hooks/useAudioPlayer";
 import { logger } from "@/lib/logger-client";
 import { useToast } from "@/components/ui/use-toast";
 import { apiClient, ApiError } from "@/lib/api/client";
-import { API_ENDPOINTS } from "@/lib/api/api-config";
+import { API_ENDPOINTS } from "@/lib/constants/api-config";
 import type { Song, Library, PlaylistOption } from "@/types/models";
 
 export default function LibraryDetailPage() {
