+++ frontend/src/pages/PlaylistsPage.tsx
@@ -16,7 +16,7 @@ import { PlaylistPlayButton } from "@/components/features/playlist-play-button";
 import { logger } from "@/lib/logger-client";
 import { useToast } from "@/components/ui/use-toast";
 import { apiClient, ApiError } from "@/lib/api/client";
-import { API_ENDPOINTS } from "@/lib/api/api-config";
+import { API_ENDPOINTS } from "@/lib/constants/api-config";
 import type { Playlist } from "@/types/models";
 
 export default function PlaylistsPage() {
